+++ src/test/resources/expected-output/java/model/common/EmptyObjectV1Dto.java
@@ -1,0 +1,9 @@
+package io.github.torand.test.model.common;
+
+import org.eclipse.microprofile.openapi.annotations.media.Schema;
+
+@Schema(name = "EmptyObjectV1", description = "TBD")
+public record EmptyObjectV1Dto (
+) {
+
+}